Update databasehw1.cpp
worked on adding an entry and creating database. adding entry works and it updates # of total records BUT you it can't find new record entered when searched. adding new Database doesn't write to the file, might be because I changed ofstream to fstream...idk why it stopped writing to the file(??)